Update from stable (#16)
* V3 Stable (#11)

* Initial commit

* C to C++ syntax in IO memories

* Version 2
* memories to vectors
* search sorting
* special characters removing in vocablurary
* trimming to print memories
* add run script
* new main architecture
* add coloring

* Fix only cyrillic words in vocablurary

* /edit command added.
* Refactor
* Cosine similarity euclidian distance //
* To lower string.

* tf-idf implemented; not used yet; looking for improvements

* slowly rewriting

* basic IO working

* to string methods done

* barebones functional here, no IO used, no search used

* IO fixed

* playing around

* llama and /search added

* /help added

* Help reformatted

* cleanup

* hotfix: search formatting and filtering

* hotfix: search using wrong vocablurary

* add repetition functions

* add cmake and update readme

* refactor to headers
