+++ README.md
@@ -2,10 +2,6 @@
 
 **Memo** is a stupidly simple note-taking app for me. Its key idea is to prevent users from viewing all their memories at once and instead to search them through keywords or exact matches. The key point is to avoid getting overwhelmed by a lot of facts or information but still be able to remember something in a small portion.
 
-### State
-
-As of now, the only possibility is to add memories and search through them. In the near future, a tag system will be added to filter the results, and there will be an ability to link memories like graphs.
-
 ### Motivation
 
 I started this project because:
@@ -14,21 +10,6 @@ I started this project because:
 - I want to practice modern C++, since I have experience primarily with C, and switching to C++ is a bit tough without practice.
 
 Since this app serves my own purpose for taking notes, you can expect me to maintain the repository over time.
-
-### Thoughts 
-As soon as simply writing notes does not really solves the problem of forgetting everything
-I want to stir the ship: using spaced repetition techniques to actually remember stuff!
-
-For that i need more data in each memory:
-  - vector<time_t> viewed; // when memory was viewed. 
-
-So, if there is linked memories to some other, I can show them and prompt user
- to recall content of desired one. 
-
-It would be as separate mode, but sounds kinda interesting, allowing the app to be used as learning platform too. But not as effective as "Anki".
-
-Also, the interesting part is analysis of notes, statistics:
- - repeated words, tags, most viewed notes, defining about what user actually thinks the most 
 
 ### Disclaimer
 
@@ -39,20 +20,45 @@ The code is messy and buggy. Use at your own risk. All of us have a life outside
 **Compile:**
 
 ```bash
-chmod +x ./build.sh
-./build.sh
+mkdir build && cd build
+cmake ..
+make
 ```
 
 **Run:**
 
 ```bash
-chmod +x ./run.sh
-./run.sh
+cd build
+chmod +x memo
+./memo
 ```
 
 **Commands:**
 
-- `/add <text>`: Adds a note with the specified text.
-- `/remove <id>`: Removes a note by ID.
-- `/search <text>`: Searches through existing memories.
-- `/exit`: Exits the program.
+- **/add <text>** (or **/a**)
+  - Add a new memory. `<text>` can have tags (e.g., `#tag`) used for filtering in search.
+
+- **/repetition** (or **/i**)
+  - Not implemented yet.
+
+- **/search <text>** (or **/s**)
+  - Search memories using `<text>`. You can filter by tags like `#tag`.
+
+- **/clear** (or **/c**)
+  - Clear the screen.
+
+- **/help** (or **/h**)
+  - Display this help message.
+
+- **/exit** (or **/e**)
+  - Exit the program.
+
+- **/remove <id>** (or **/r**)
+  - Remove a memory with the specified `<id>`.
+
+- **/update <id> <text>** (or **/u**)
+  - Update a memory with the specified `<id>` to the new `<text>`.
+
+- **/all**
+  - Display all memories.
+
